Add setpgid for all called commands
Need to make sure signals are sent to all processes as sub-shells
interfere with signal propagation. Using Setpgid to create a process
group for the command and all it's subprocesses fixes this by allowing
child to send the signals to the pgid, which thus gets sent to all
processes in that group.

This is required for upcoming change to use sub-shells for all executed
commands to eliminate need for shell parsing.

With this change it is always set to true and the pgid is always used to
send signals to the process(es).

In the future it could be made configurable if needed.